Compute the objective function value

Compute the objective function value

Usage

compute_ofv(
  eta,
  qmod,
  sigma,
  omega_inv,
  all_cmt,
  log_transformation,
  lambda = 1,
  idvaliddata,
  idDV,
  idcmt,
  idblq = NULL,
  idlloq = NULL,
  ...
)

do_compute_ofv(eta, argofv, ...)

Arguments

eta

a named vector/list of parameters

qmod, sigma, log_transformation, omega_inv, all_cmt, lambda

generated by preprocess.ofv.fix

idvaliddata, idDV, idcmt

generated by preprocess.ofv.id

idblq, idlloq

optionally generated by preprocess.ofv.id

...

for compatibility (not used)

argofv

above mentioned arguments as a list

Value

a single numeric value (the objective function value)

Details

This function is called iteratively by the optimization function. Arguments should not be passed directly, but generated by the pre-processing functions (see preprocess.ofv).