Simulate with mapbayr

A wrapper around mrgsolve::mrgsim() for results generated from mapbayest(). Exported for the purpose of utility but might be prone to changes.

Usage

do_mapbayr_sim(
  x,
  data,
  recsort = 3,
  output = "df",
  ...,
  eta = NULL,
  nrep = NULL,
  new_omega = NULL,
  new_sigma = NULL
)

Arguments

x

the model object

data

NMTRAN-like data set

recsort

record sorting flag. Defaulted to 3. See mrgsolve::mrgsim().

output

type of object returned. Defaulted to "df" for a data.frame. See mrgsolve::mrgsim().

...

passed to mrgsolve::mrgsim().

eta

a matrix of individual point estimates of ETA. Most likely obtained with get_eta().

nrep

number of replicates. If used, the original "ID" in the data will be replaced by unique identifiers.

new_omega, new_sigma

New "omega" and "sigma" matrices to use instead of those defined in "$OMEGA" and "$SIGMA".

Value

An output from mrgsolve::mrgsim().

Examples

library(mrgsolve)
mod <- exmodel(1, exdata = FALSE)
dat <- exdata(ID = c(1,2))

# Classic framework
set.seed(123)
do_mapbayr_sim(x = mod, data = dat, Request = "DV")
#>   ID time        DV
#> 1  1  0.0   0.00000
#> 2  1  1.5 137.02297
#> 3  1  4.4 142.99304
#> 4  1  7.1 144.53870
#> 5  1 24.6  67.94558
#> 6  2  0.0   0.00000
#> 7  2 25.9  33.72063

# No random effect
do_mapbayr_sim(x = zero_re(mod), data = dat)
#>   ID time        DEPOT  CENTRAL        DV
#> 1  1  0.0 1.000000e+04    0.000   0.00000
#> 2  1  1.5 2.231302e+03 7368.309 105.26156
#> 3  1  4.4 1.227734e+02 8118.003 115.97146
#> 4  1  7.1 8.251050e+00 7060.194 100.85992
#> 5  1 24.6 2.073399e-07 2600.520  37.15028
#> 6  2  0.0 1.000000e+04    0.000   0.00000
#> 7  2 25.9 5.666652e-08 2414.339  34.49056
do_mapbayr_sim(x = mod, data = dat, new_omega = "zero_re")
#>   ID time        DEPOT  CENTRAL        DV
#> 1  1  0.0 1.000000e+04    0.000   0.00000
#> 2  1  1.5 2.231302e+03 7368.309 117.10302
#> 3  1  4.4 1.227734e+02 8118.003 129.11558
#> 4  1  7.1 8.251050e+00 7060.194 132.40069
#> 5  1 24.6 2.073399e-07 2600.520  34.77546
#> 6  2  0.0 1.000000e+04    0.000   0.00000
#> 7  2 25.9 5.666652e-08 2414.339  39.71998

# New random effects
## New omega matrix
do_mapbayr_sim(x = mod, data = dat, new_omega = dmat(0.1, 0.03, 0.01), nrep = 10)
#>    ID time        DEPOT   CENTRAL         DV
#> 1   1  0.0 1.000000e+04    0.0000   0.000000
#> 2   1  1.5 2.276094e+03 6804.0806 265.331815
#> 3   1  4.4 1.301442e+02 6177.1725 158.728030
#> 4   1  7.1 9.064991e+00 4345.9003 118.077947
#> 5   1 24.6 2.869783e-07  387.2795  13.522088
#> 6   2  0.0 1.000000e+04    0.0000   0.000000
#> 7   2 25.9 2.141098e-06 4648.6116  56.019397
#> 8   3  0.0 1.000000e+04    0.0000   0.000000
#> 9   3  1.5 2.762745e+03 6989.9395  56.135306
#> 10  3  4.4 2.297557e+02 8598.0402  88.600588
#> 11  3  7.1 2.268191e+01 7942.2810  86.034802
#> 12  3 24.6 6.886618e-06 4061.2451  54.117440
#> 13  4  0.0 1.000000e+04    0.0000   0.000000
#> 14  4 25.9 1.036727e-09 2516.7885  52.007899
#> 15  5  0.0 1.000000e+04    0.0000   0.000000
#> 16  5  1.5 2.334198e+03 7358.5978 152.583347
#> 17  5  4.4 1.401319e+02 8474.2596  91.850326
#> 18  5  7.1 1.021374e+01 7637.2710 120.184071
#> 19  5 24.6 4.335280e-07 3518.8557  39.819529
#> 20  6  0.0 1.000000e+04    0.0000   0.000000
#> 21  6 25.9 4.139697e-07 2327.8765  29.319303
#> 22  7  0.0 1.000000e+04    0.0000   0.000000
#> 23  7  1.5 2.633200e+03 6960.7802  83.970767
#> 24  7  4.4 1.995658e+02 7954.2999  93.251168
#> 25  7  7.1 1.806996e+01 6880.5089 112.721108
#> 26  7 24.6 3.132786e-06 2310.0333  21.769806
#> 27  8  0.0 1.000000e+04    0.0000   0.000000
#> 28  8 25.9 7.249688e-09 2072.4060  33.189774
#> 29  9  0.0 1.000000e+04    0.0000   0.000000
#> 30  9  1.5 2.364118e+03 7188.2189  83.245338
#> 31  9  4.4 1.454665e+02 7881.1208  96.827111
#> 32  9  7.1 1.084845e+01 6718.0997 104.792134
#> 33  9 24.6 5.341592e-07 2129.1815  27.472297
#> 34 10  0.0 1.000000e+04    0.0000   0.000000
#> 35 10 25.9 1.283792e-11 1348.5991  20.993151
#> 36 11  0.0 1.000000e+04    0.0000   0.000000
#> 37 11  1.5 2.545595e+03 7023.1075  74.393243
#> 38 11  4.4 1.807099e+02 7882.0683  85.818048
#> 39 11  7.1 1.539582e+01 6751.4651  92.033760
#> 40 11 24.6 1.798545e-06 2152.0885  32.268981
#> 41 12  0.0 1.000000e+04    0.0000   0.000000
#> 42 12 25.9 8.323596e-08  355.2793   7.171505
#> 43 13  0.0 1.000000e+04    0.0000   0.000000
#> 44 13  1.5 2.529678e+03 7045.8704  93.523065
#> 45 13  4.4 1.774154e+02 7916.2657  91.494866
#> 46 13  7.1 1.494544e+01 6800.7034  95.263939
#> 47 13 24.6 1.622716e-06 2216.2989  32.249572
#> 48 14  0.0 1.000000e+04    0.0000   0.000000
#> 49 14 25.9 2.183283e-08 1781.7069  23.221106
#> 50 15  0.0 1.000000e+04    0.0000   0.000000
#> 51 15  1.5 2.670168e+03 6946.0931  79.566594
#> 52 15  4.4 2.078962e+02 8034.3471 103.015232
#> 53 15  7.1 1.930262e+01 7013.0763  84.420295
#> 54 15 24.6 3.937631e-06 2487.4881  34.088192
#> 55 16  0.0 1.000000e+04    0.0000   0.000000
#> 56 16 25.9 1.729749e-09 2035.4942  28.515754
#> 57 17  0.0 1.000000e+04    0.0000   0.000000
#> 58 17  1.5 2.424954e+03 7347.6667 120.304262
#> 59 17  4.4 1.567222e+02 8793.9428 148.691480
#> 60 17  7.1 1.223480e+01 8176.2887 134.198399
#> 61 17 24.6 8.107715e-07 4582.2766  97.827273
#> 62 18  0.0 1.000000e+04    0.0000   0.000000
#> 63 18 25.9 6.518038e-07 3723.3268  15.571808
#> 64 19  0.0 1.000000e+04    0.0000   0.000000
#> 65 19  1.5 2.192477e+03 7436.1270  39.351482
#> 66 19  4.4 1.166119e+02 8248.8953  75.164909
#> 67 19  7.1 7.593219e+00 7257.4767  80.971182
#> 68 19 24.6 1.555237e-07 2899.1595  30.137121
#> 69 20  0.0 1.000000e+04    0.0000   0.000000
#> 70 20 25.9 8.703064e-08 1901.7247  46.053786

## Matrix with "eta" as mean and "new_omega" as variance covariance matrix
etamat <- get_eta(est001, output = "num")[1:2,]

do_mapbayr_sim(
  x = mod, data = dat, nrep = 10,
  eta = etamat, new_omega = dmat(0.1, 0.03, 0.01)
)
#>    ID time        DEPOT   CENTRAL         DV
#> 1   1  0.0 1.000000e+04    0.0000   0.000000
#> 2   1  1.5 1.793532e+03 7568.3219 119.354688
#> 3   1  4.4 6.469637e+01 7351.0008 143.715406
#> 4   1  7.1 2.934649e+00 5889.5959  85.391901
#> 5   1 24.6 6.057606e-09 1322.1415  29.196512
#> 6   2  0.0 1.000000e+04    0.0000   0.000000
#> 7   2 25.9 2.283347e-07  655.7038  10.851148
#> 8   3  0.0 1.000000e+04    0.0000   0.000000
#> 9   3  1.5 2.117685e+03 7570.1252 125.287126
#> 10  3  4.4 1.053238e+02 8515.9439 167.867253
#> 11  3  7.1 6.442831e+00 7667.1534 126.734090
#> 12  3 24.6 8.813555e-08 3596.1156  69.411687
#> 13  4  0.0 1.000000e+04    0.0000   0.000000
#> 14  4 25.9 5.618487e-09 1691.4945  21.886587
#> 15  5  0.0 1.000000e+04    0.0000   0.000000
#> 16  5  1.5 1.748034e+03 7665.7367 134.336723
#> 17  5  4.4 5.999935e+01 7555.2601 128.637467
#> 18  5  7.1 2.598582e+00 6178.8610  99.572911
#> 19  5 24.6 4.425286e-09 1593.7674  44.641610
#> 20  6  0.0 1.000000e+04    0.0000   0.000000
#> 21  6 25.9 2.055474e-07 2704.4301  64.932326
#> 22  7  0.0 1.000000e+04    0.0000   0.000000
#> 23  7  1.5 2.249374e+03 6841.3859 134.008880
#> 24  7  4.4 1.257132e+02 6224.5774  91.550249
#> 25  7  7.1 8.572190e+00 4404.0178  51.186653
#> 26  7 24.6 2.364479e-07  409.4177   7.944375
#> 27  8  0.0 1.000000e+04    0.0000   0.000000
#> 28  8 25.9 2.024871e-08 2194.1030  34.479422
#> 29  9  0.0 1.000000e+04    0.0000   0.000000
#> 30  9  1.5 2.072071e+03 7053.2729 122.907117
#> 31  9  4.4 9.880687e+01 6417.8915 136.313006
#> 32  9  7.1 5.811867e+00 4641.6989  50.699845
#> 33  9 24.6 6.149278e-08  513.3733   8.071178
#> 34 10  0.0 1.000000e+04    0.0000   0.000000
#> 35 10 25.9 1.239100e-08 1863.2347  19.624588
#> 36 11  0.0 1.000000e+04    0.0000   0.000000
#> 37 11  1.5 1.737039e+03 7371.0028 106.951572
#> 38 11  4.4 5.889900e+01 6498.5237 117.447730
#> 39 11  7.1 2.522116e+00 4739.2870  72.210033
#> 40 11 24.6 2.976301e-09  576.3854  10.021807
#> 41 12  0.0 1.000000e+04    0.0000   0.000000
#> 42 12 25.9 2.317578e-06 2106.4043  30.294332
#> 43 13  0.0 1.000000e+04    0.0000   0.000000
#> 44 13  1.5 1.828884e+03 7542.5852  75.605272
#> 45 13  4.4 6.850874e+01 7373.6094  97.710783
#> 46 13  7.1 3.218704e+00 5922.9161  94.232172
#> 47 13 24.6 8.094597e-09 1347.8501  17.929766
#> 48 14  0.0 1.000000e+04    0.0000   0.000000
#> 49 14 25.9 9.486998e-09  547.0377  12.609039
#> 50 15  0.0 1.000000e+04    0.0000   0.000000
#> 51 15  1.5 2.664677e+03 6770.1071 110.580542
#> 52 15  4.4 2.066447e+02 7293.7041 127.620902
#> 53 15  7.1 1.911545e+01 5898.3858  95.141362
#> 54 15 24.6 3.806940e-06 1251.4617  26.109768
#> 55 16  0.0 1.000000e+04    0.0000   0.000000
#> 56 16 25.9 8.145580e-08 1335.9604  12.006268
#> 57 17  0.0 1.000000e+04    0.0000   0.000000
#> 58 17  1.5 1.568652e+03 7597.8521 152.336690
#> 59 17  4.4 4.367283e+01 6756.9768 110.732688
#> 60 17  7.1 1.556536e+00 5077.0225 123.855190
#> 61 17 24.6 6.572205e-10  762.4210  12.581644
#> 62 18  0.0 1.000000e+04    0.0000   0.000000
#> 63 18 25.9 4.878177e-09 1128.0833  23.508400
#> 64 19  0.0 1.000000e+04    0.0000   0.000000
#> 65 19  1.5 1.984133e+03 7521.9678 121.428969
#> 66 19  4.4 8.700439e+01 7826.2201 149.507273
#> 67 19  7.1 4.733351e+00 6589.6346 116.630795
#> 68 19 24.6 3.033609e-08 2013.9499  28.049690
#> 69 20  0.0 1.000000e+04    0.0000   0.000000
#> 70 20 25.9 5.361778e-08 2529.5491  43.210817